+++ fixtures/gists-app/app/routes/gists.js
@@ -1,5 +1,5 @@
 import { Outlet } from "react-router-dom";
-import { json } from "@remix-run/express";
+import { json } from "@remix-run/node";
 import { Link, useRouteData, usePendingLocation } from "@remix-run/react";
 
 import styles from "css:../styles/gists.css";