+++ fixtures/gists-app/app/routes/methods.tsx
@@ -1,6 +1,6 @@
 import * as React from "react";
-import type { LoaderFunction, ActionFunction } from "@remix-run/express";
-import { json, redirect } from "@remix-run/express";
+import type { LoaderFunction, ActionFunction } from "@remix-run/node";
+import { json, redirect } from "@remix-run/node";
 import type { FormProps } from "@remix-run/react";
 import { useRouteData, Form, usePendingFormSubmit } from "@remix-run/react";