+++ fixtures/gists-app/app/routes/prefs.tsx
@@ -3,8 +3,8 @@ import type {
   MetaFunction,
   LoaderFunction,
   ActionFunction
-} from "@remix-run/express";
-import { redirect } from "@remix-run/express";
+} from "@remix-run/node";
+import { redirect } from "@remix-run/node";
 import { useRouteData, Form, useSubmit } from "@remix-run/react";
 
 import { userPrefsCookie } from "../cookies";