+++ fixtures/tutorial/app/routes/gists.tsx
@@ -1,5 +1,5 @@
-import type { LoaderFunction, ActionFunction } from "@remix-run/express";
-import { redirect } from "@remix-run/express";
+import type { LoaderFunction, ActionFunction } from "@remix-run/node";
+import { redirect } from "@remix-run/node";
 import { useRouteData } from "@remix-run/react";
 
 export let loader: LoaderFunction = () => {