+++ fixtures/tutorial/app/routes/team/$member.tsx
@@ -1,4 +1,4 @@
-import type { LoaderFunction } from "@remix-run/express";
+import type { LoaderFunction } from "@remix-run/node";
 import { useRouteData } from "@remix-run/react";
 
 interface User {