+++ packages/vercel/index.ts
@@ -1,46 +1,4 @@
-export type {
-  // fetch
-  HeadersInit,
-  RequestInit,
-  RequestInfo,
-  ResponseInit,
-  // cookies & sessions
-  Cookie,
-  CookieOptions,
-  SessionData,
-  Session,
-  SessionStorage,
-  SessionIdStorageStrategy,
-  // route modules
-  RouteComponent,
-  ErrorBoundaryComponent,
-  HeadersFunction,
-  MetaFunction,
-  LinksFunction,
-  ActionFunction,
-  ActionFunction as Action, // shorthand
-  LoaderFunction,
-  LoaderFunction as Loader // shorthand
-} from "@remix-run/node";
-export {
-  // fetch
-  Headers,
-  Request,
-  Response,
-  fetch,
-  // cookies & sessions
-  createCookie,
-  isCookie,
-  createSession,
-  isSession,
-  createSessionStorage,
-  createCookieSessionStorage,
-  createFileSessionStorage,
-  createMemorySessionStorage,
-  // responses
-  json,
-  redirect
-} from "@remix-run/node";
+import "./globals";
 
 export type { GetLoadContextFunction, RequestHandler } from "./server";
 export { createRequestHandler } from "./server";